Make discord linking an option instead of a requirement

+++ src/main/java/com/github/riku32/discordlink/discord/listeners/CrosschatListener.java
@@ -34,44 +34,57 @@ public class CrosschatListener extends ListenerAdapter {
         if (!event.getGuild().getId().equals(plugin.getBot().getGuild().getId())) return;
         if (!event.getChannel().getId().equals(plugin.getBot().getChannel().getId())) return;
 
+        Member member = Objects.requireNonNull(event.getMember());
+
         Optional<PlayerInfo> optionalPlayerInfo = plugin.getDatabase().getPlayerInfo(event.getAuthor().getId());
-        if (!optionalPlayerInfo.isPresent()) {
-            event.getAuthor().openPrivateChannel().queue(privateChannel -> {
-                privateChannel.sendMessage(new EmbedBuilder()
+        if (plugin.getPluginConfig().isLinkRequired()) {
+            if (!optionalPlayerInfo.isPresent()) {
+                event.getAuthor().openPrivateChannel().queue(privateChannel -> privateChannel.sendMessage(new EmbedBuilder()
                         .setColor(Constants.Colors.FAIL)
                         .setTitle("Not linked")
                         .setDescription("Your can't use crosschat unless you link your Minecraft account. " +
                                 "Please join the server and type `/link` to start the link process.")
                         .build())
-                        .queue();
-            });
-            event.getMessage().delete().queue();
+                        .queue());
+                event.getMessage().delete().queue();
+                return;
+            }
+
+            if (!optionalPlayerInfo.get().isVerified()) {
+                event.getAuthor().openPrivateChannel().queue(privateChannel -> {
+                    try {
+                        privateChannel.retrieveMessageById(plugin.getDatabase().getMessageId(optionalPlayerInfo.get().getDiscordID()))
+                                .queue(verificationMessage -> privateChannel.sendMessage(new EmbedBuilder()
+                                    .setColor(Constants.Colors.FAIL)
+                                    .setTitle("Not verified")
+                                    .setDescription(String.format("Your can't use crosschat unless you verify your Minecraft account link. " +
+                                            "Please press either the verify or cancel buttons on the [verification message](%s).", verificationMessage.getJumpUrl()))
+                                    .build())
+                                    .queue());
+                    } catch (SQLException exception) {
+                        exception.printStackTrace();
+                    }
+                });
+                event.getMessage().delete().queue();
+                return;
+            }
+        } else if (!optionalPlayerInfo.isPresent() || !optionalPlayerInfo.get().isVerified()) {
+            Bukkit.broadcastMessage(ChatColor.translateAlternateColorCodes('&',
+                    plugin.getPluginConfig().getDiscordFormatUnlinked()
+                            .replaceAll("%color%", Util.colorToChatString(
+                                    member.getColor() == null ? ChatColor.GRAY.getColor() : member.getColor()))
+                            .replaceAll("%tag%", member.getUser().getAsTag()))
+                    // Message is replaced after colorized so message does not colorize
+                    .replaceAll("%message%",
+                            event.getMessage().getContentStripped().isEmpty()
+                                    ? ChatColor.DARK_GRAY + String.format("%d attachments", event.getMessage().getAttachments().size())
+                                    : event.getMessage().getContentStripped())
+                    // Minecraft chat doesnt allow newlines so all discord message newlines will be removed
+                    .replaceAll("\n", ""));
             return;
         }
 
         PlayerInfo playerInfo = optionalPlayerInfo.get();
-
-        if (!playerInfo.isVerified()) {
-            event.getAuthor().openPrivateChannel().queue(privateChannel -> {
-                try {
-                    privateChannel.retrieveMessageById(plugin.getDatabase().getMessageId(playerInfo.getDiscordID())).queue(verificationMessage -> {
-                        privateChannel.sendMessage(new EmbedBuilder()
-                                .setColor(Constants.Colors.FAIL)
-                                .setTitle("Not verified")
-                                .setDescription(String.format("Your can't use crosschat unless you verify your Minecraft account link. " +
-                                        "Please press either the verify or cancel buttons on the [verification message](%s).", verificationMessage.getJumpUrl()))
-                                .build())
-                                .queue();
-                    });
-                } catch (SQLException exception) {
-                    exception.printStackTrace();
-                }
-            });
-            event.getMessage().delete().queue();
-            return;
-        }
-
-        Member member = Objects.requireNonNull(event.getMember());
 
         OfflinePlayer offlinePlayer = plugin.getServer().getOfflinePlayer(playerInfo.getUuid());
         if (!offlinePlayer.hasPlayedBefore()) {
@@ -81,7 +94,7 @@ public class CrosschatListener extends ListenerAdapter {
         }
 
         Bukkit.broadcastMessage(ChatColor.translateAlternateColorCodes('&',
-                plugin.getPluginConfig().getDiscordFormat()
+                plugin.getPluginConfig().getDiscordFormatLinked()
                         .replaceAll("%color%", Util.colorToChatString(
                                 member.getColor() == null ? ChatColor.GRAY.getColor() : member.getColor()))
                         .replaceAll("%username%", Objects.requireNonNull(offlinePlayer.getName()))
